// BannerClock/src/Interfaces/ISettings.ts
export interface ISettings{
    digitalOnly: boolean;
    hour12: boolean;
    displayDay: boolean;
    spList: string;
    ClassicBannerId: string;
    ClassicFooterId: string;
    ModernBannerComId: string;
    ModernFooterComId: string;
    ModernBannerTeamId: string;
    ModernFooterTeamId: string;
}